% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_limits.R
\docType{data}
\name{AXIS_PADDING_FRACTION}
\alias{AXIS_PADDING_FRACTION}
\title{Default axis padding fraction}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
AXIS_PADDING_FRACTION
}
\description{
The fraction of the data range to add as padding around axis limits.
Default is 0.025 (2.5% padding on each side).
}
\keyword{internal}
