% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{exclude_transects}
\alias{exclude_transects}
\title{Exclude specific transects}
\usage{
exclude_transects(filter, transects)
}
\arguments{
\item{filter}{data_filter object}

\item{transects}{Character vector. Transect IDs to exclude}
}
\value{
Modified data_filter object
}
\description{
Exclude specific transects
}
\examples{
filter <- create_data_filter() |>
  exclude_transects(c("PAIS_T001", "GUIS_T015"))
}
