% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{get_transects_by_park}
\alias{get_transects_by_park}
\title{Get transects for a specific park}
\usage{
get_transects_by_park(data, park)
}
\arguments{
\item{data}{Dataframe with transect data}

\item{park}{Character. Park code}
}
\value{
Character vector of transect IDs
}
\description{
Get transects for a specific park
}
\examples{
\dontrun{
pais_transects <- get_transects_by_park(data, "PAIS")
}
}
