% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{get_transects_by_year}
\alias{get_transects_by_year}
\title{Get transects measured in a specific year}
\usage{
get_transects_by_year(data, year)
}
\arguments{
\item{data}{Dataframe with transect data}

\item{year}{Numeric. Year}
}
\value{
Character vector of transect IDs
}
\description{
Get transects measured in a specific year
}
\examples{
\dontrun{
transects_2020 <- get_transects_by_year(data, 2020)
}
}
