% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{get_years_by_transect}
\alias{get_years_by_transect}
\title{Get years for a specific transect}
\usage{
get_years_by_transect(data, transect)
}
\arguments{
\item{data}{Dataframe with transect data}

\item{transect}{Character. Transect ID}
}
\value{
Numeric vector of years
}
\description{
Get years for a specific transect
}
\examples{
\dontrun{
years <- get_years_by_transect(data, "PAIS_T001")
}
}
