% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GULNcoastal-package.R
\docType{package}
\name{GULNcoastal-package}
\alias{GULNcoastal}
\alias{GULNcoastal-package}
\title{GULNcoastal: Gulf Coast Network Coastal Monitoring Analysis Tools}
\description{
Tools and functions for analyzing coastal monitoring data from Gulf Coast Network park units. This package provides data processing, analysis, and visualization capabilities for coastal ecosystem monitoring.
}
\author{
\strong{Maintainer}: Alisa Glenewinkel-Meyer \email{alisa.glenewinkel@gmail.com}

}
\keyword{internal}
