% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{is_data_filter}
\alias{is_data_filter}
\title{Check if object is a data_filter}
\usage{
is_data_filter(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if object is a data_filter
}
