% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_config.R
\name{load_config}
\alias{load_config}
\title{Load configuration from file}
\usage{
load_config(filepath)
}
\arguments{
\item{filepath}{Character. Path to configuration file (RDS format)}
}
\value{
plot_config object
}
\description{
Load configuration from file
}
