% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_config.R
\name{modify_config}
\alias{modify_config}
\title{Merge configuration with override parameters}
\usage{
modify_config(config, ...)
}
\arguments{
\item{config}{plot_config object. Base configuration}

\item{...}{Named parameters to override}
}
\value{
Modified plot_config object
}
\description{
Takes a base configuration and updates specific parameters.
Useful for applying minor modifications to preset configs.
}
