% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{print.data_filter}
\alias{print.data_filter}
\title{Print data filter summary}
\usage{
\method{print}{data_filter}(x, ...)
}
\arguments{
\item{x}{data_filter object}

\item{...}{Additional arguments (unused)}
}
\description{
Print data filter summary
}
