% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_config.R
\name{print.plot_config}
\alias{print.plot_config}
\title{Print configuration summary}
\usage{
\method{print}{plot_config}(x, ...)
}
\arguments{
\item{x}{plot_config object}

\item{...}{Additional arguments (unused)}
}
\description{
Print configuration summary
}
