% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{print.resolved_filter}
\alias{print.resolved_filter}
\title{Print resolved filter summary}
\usage{
\method{print}{resolved_filter}(x, ...)
}
\arguments{
\item{x}{resolved_filter object}

\item{...}{Additional arguments (unused)}
}
\description{
Print resolved filter summary
}
