% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_config.R
\name{print.validation_report}
\alias{print.validation_report}
\title{Print validation report}
\usage{
\method{print}{validation_report}(x, ...)
}
\arguments{
\item{x}{validation_report object}

\item{...}{Additional arguments (unused)}
}
\description{
Print validation report
}
