% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_config.R
\name{save_config}
\alias{save_config}
\title{Save configuration to file}
\usage{
save_config(config, filepath)
}
\arguments{
\item{config}{plot_config object}

\item{filepath}{Character. Path to save configuration (RDS format)}
}
\description{
Save configuration to file
}
