% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{set_name}
\alias{set_name}
\title{Set filter name}
\usage{
set_name(filter, name)
}
\arguments{
\item{filter}{data_filter object}

\item{name}{Character. Filter name}
}
\value{
Modified data_filter object
}
\description{
Set filter name
}
\examples{
filter <- create_data_filter() |>
  set_name("baseline_comparison")
}
