% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{set_parks}
\alias{set_parks}
\title{Set park filter}
\usage{
set_parks(filter, parks)
}
\arguments{
\item{filter}{data_filter object}

\item{parks}{Character vector. Park codes to include}
}
\value{
Modified data_filter object
}
\description{
Set park filter
}
\examples{
filter <- create_data_filter() |>
  set_parks(c("PAIS", "GUIS"))
}
