% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{set_transects}
\alias{set_transects}
\title{Set transect filter}
\usage{
set_transects(filter, transects)
}
\arguments{
\item{filter}{data_filter object}

\item{transects}{Character vector. Transect IDs to include}
}
\value{
Modified data_filter object
}
\description{
Set transect filter
}
\examples{
filter <- create_data_filter() |>
  set_transects(c("PAIS_T001", "PAIS_T002"))
}
