% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter.R
\name{set_years}
\alias{set_years}
\title{Set year selector for filter}
\usage{
set_years(filter, selector)
}
\arguments{
\item{filter}{data_filter object}

\item{selector}{year_selector object}
}
\value{
Modified data_filter object
}
\description{
Set year selector for filter
}
\examples{
filter <- create_data_filter() |>
  set_years(years_recent(3))
}
