% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colors.R
\docType{data}
\name{YEAR_CADENCE}
\alias{YEAR_CADENCE}
\title{Expected cadence between survey years}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
YEAR_CADENCE
}
\description{
The typical interval (in years) between successive surveys. Used by
\code{\link{format_year_sequence}} to identify consecutive runs of years.
Default is 2 years.
}
\keyword{internal}
