% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_selectors.R
\name{years_combine}
\alias{years_combine}
\title{Combine multiple year selectors}
\usage{
years_combine(...)
}
\arguments{
\item{...}{\code{year_selector} objects to combine}
}
\value{
A \code{year_selector} object
}
\description{
Creates a selector that unions the results of multiple selectors.
Each selector is applied to the park's years, and results are combined.
}
\examples{
# First year + years since 2018
combined <- years_combine(
  years_first(1),
  years_since(2018)
)
}
\seealso{
\code{\link{list_year_selectors}} for all available selectors
\code{\link{preview_year_selection}} to test selector before use
\code{\link{create_data_filter}} to use with data filtering
}
