% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_selectors.R
\name{years_since}
\alias{years_since}
\title{Select years since a given year}
\usage{
years_since(year)
}
\arguments{
\item{year}{Numeric. Start year (inclusive)}
}
\value{
A \code{year_selector} object
}
\description{
Returns all years greater than or equal to the specified year.
Useful for filtering to recent time periods.
}
\examples{
# All surveys from 2018 onward
selector <- years_since(2018)
}
