% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_selectors.R
\name{years_until}
\alias{years_until}
\title{Select years up to a given year}
\usage{
years_until(year)
}
\arguments{
\item{year}{Numeric. End year (inclusive)}
}
\value{
A \code{year_selector} object
}
\description{
Returns all years less than or equal to the specified year.
Useful for historical analysis.
}
\examples{
# All surveys up to 2015
selector <- years_until(2015)
}
